# eventhub/packages/contracts/openapi.yaml
openapi: 3.1.0

info:
  title: EventHub API
  version: 0.1.0
  description: Портал продажу квитків на події

servers:
  - url: http://localhost:4010
    description: Мок-сервер Prism
  - url: http://localhost:3000
    description: Локальний api-gateway

paths:
  /events:
    get:
      operationId: listEvents
      summary: Список подій із курсорною пагінацією
      tags: [catalog]
      parameters:
        - name: city
          in: query
          required: false
          schema:
            type: string
            examples: [Вінниця]
        - name: category
          in: query
          required: false
          description: Категорія події
          schema:
            type: string
            examples: [concert]
        - name: from
          in: query
          required: false
          description: Показати події, що починаються не раніше цієї дати
          schema: { type: string, format: date }
        - name: to
          in: query
          required: false
          description: Показати події, що починаються не пізніше цієї дати
          schema: { type: string, format: date }
        - name: cursor
          in: query
          required: false
          description: Непрозорий курсор із попередньої відповіді
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: Сторінка подій
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventPage"
              examples:
                filled:
                  summary: Сторінка з подіями
                  value:
                    items:
                      - id: 018f2c1e-4a7b-7c3d-9e1f-2a5b6c7d8e9f
                        title: Симфонічний вечір
                        category: concert
                        status: scheduled
                        startsAt: "2026-10-12T19:00:00Z"
                        endsAt: "2026-10-12T21:30:00Z"
                        venue:
                          id: 018f2c1e-0000-7c3d-9e1f-2a5b6c7d8e00
                          name: Філармонія
                          city: Вінниця
                        minPrice: { amount: 25000, currency: UAH }
                        availableSeats: 143
                      - id: 018f2c1e-4a7b-7c3d-9e1f-2a5b6c7d8ea0
                        title: Джазовий квартет
                        category: concert
                        status: scheduled
                        startsAt: "2026-10-19T20:00:00Z"
                        venue:
                          id: 018f2c1e-0000-7c3d-9e1f-2a5b6c7d8e01
                          name: Клуб «Підвал»
                          city: Вінниця
                        minPrice: { amount: 18000, currency: UAH }
                        availableSeats: 12
                    nextCursor: "MjAyNi0xMC0xOVQyMDowMDowMFo"
                empty:
                  summary: Порожній каталог
                  value:
                    items: []
                    nextCursor: null
        "400":
          $ref: "#/components/responses/BadRequest"

  /events/{eventId}:
    get:
      operationId: getEvent
      summary: Одна подія
      tags: [catalog]
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Подія
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
              examples:
                default:
                  value:
                    id: 018f2c1e-4a7b-7c3d-9e1f-2a5b6c7d8e9f
                    title: Симфонічний вечір
                    description: Програма з творів Лисенка та Дворжака
                    category: concert
                    status: scheduled
                    imageUrl: https://example.com/images/symphony-evening.jpg
                    startsAt: "2026-10-12T19:00:00Z"
                    endsAt: "2026-10-12T21:30:00Z"
                    venue:
                      id: 018f2c1e-0000-7c3d-9e1f-2a5b6c7d8e00
                      name: Філармонія
                      city: Вінниця
                    minPrice: { amount: 25000, currency: UAH }
                    availableSeats: 143
        "404":
          $ref: "#/components/responses/NotFound"

  /orders:
    post:
      operationId: createOrder
      summary: Створити замовлення на квитки
      description: >-
        Оплати ще немає (див. домашнє завдання ЛР2): замовлення одразу
        переходить у статус confirmed, місця резервуються в момент створення.
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: >-
            Замовлення створено. cancellationToken повертається лише цього
            разу - збережіть його, він потрібен для скасування.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderResponse"
              examples:
                default:
                  value:
                    id: 018f2c1e-ffff-7c3d-9e1f-000000000001
                    eventId: 018f2c1e-4a7b-7c3d-9e1f-2a5b6c7d8e9f
                    quantity: 2
                    status: confirmed
                    totalPrice: { amount: 50000, currency: UAH }
                    createdAt: "2026-10-03T12:00:00Z"
                    cancellationToken: 7f9c3b2a-1d4e-4a6b-8c2d-5e6f7a8b9c0d
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/SeatsUnavailable"
        "410":
          $ref: "#/components/responses/EventGone"

  /orders/{orderId}:
    get:
      operationId: getOrder
      summary: Переглянути замовлення
      description: >-
        orderId - непрозоре посилання з підтвердження замовлення. Самого
        знання orderId достатньо для перегляду; cancellationToken потрібен
        лише для скасування (крок 8.3 методички, принцип найменших прав).
      tags: [orders]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Замовлення
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
              examples:
                default:
                  value:
                    id: 018f2c1e-ffff-7c3d-9e1f-000000000001
                    eventId: 018f2c1e-4a7b-7c3d-9e1f-2a5b6c7d8e9f
                    quantity: 2
                    status: confirmed
                    totalPrice: { amount: 50000, currency: UAH }
                    createdAt: "2026-10-03T12:00:00Z"
        "404":
          $ref: "#/components/responses/NotFound"

  /orders/{orderId}/cancel:
    post:
      operationId: cancelOrder
      summary: Скасувати замовлення
      description: >-
        Повторний виклик на вже скасованому замовленні - безпечна (ідемпотентна)
        операція: повертає 200 із тим самим станом, а не помилку. Скасування,
        на відміну від створення, не має побічного ефекту, яким можна
        "нашкодити" повторно.
      tags: [orders]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelOrderRequest"
      responses:
        "200":
          description: Замовлення скасовано (або вже було скасоване раніше)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
              examples:
                default:
                  value:
                    id: 018f2c1e-ffff-7c3d-9e1f-000000000001
                    eventId: 018f2c1e-4a7b-7c3d-9e1f-2a5b6c7d8e9f
                    quantity: 2
                    status: cancelled
                    totalPrice: { amount: 50000, currency: UAH }
                    createdAt: "2026-10-03T12:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: >-
            Замовлення не знайдено АБО cancellationToken не збігається.
            Навмисно той самий код і те саме тіло в обох випадках: якби
            невірний токен повертав 403, а чужий orderId - 404, зловмисник
            зміг би перебором дізнатись, які orderId взагалі існують.
          $ref: "#/components/responses/NotFound"

components:
  schemas:
    Event:
      type: object
      required: [id, title, category, status, startsAt, venue, minPrice, availableSeats]
      properties:
        id:          { type: string, format: uuid }
        title:       { type: string }
        description: { type: string }
        category:    { type: string }
        imageUrl:    { type: string, format: uri }
        status:
          type: string
          enum: [scheduled, cancelled]
          description: >-
            scheduled - подія відбудеться; cancelled - організатор скасував
            подію, нові замовлення на неї неможливі (див. POST /orders, 410).
        startsAt:    { type: string, format: date-time }
        endsAt:      { type: string, format: date-time }
        venue:       { $ref: "#/components/schemas/VenueSummary" }
        minPrice:    { $ref: "#/components/schemas/Money" }
        availableSeats: { type: integer, minimum: 0 }

    VenueSummary:
      type: object
      required: [id, name, city]
      properties:
        id:   { type: string, format: uuid }
        name: { type: string }
        city: { type: string }

    Money:
      type: object
      required: [amount, currency]
      description: Сума в мінімальних одиницях валюти (копійках)
      properties:
        amount:   { type: integer }
        currency: { type: string, enum: [UAH] }

    EventPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Event" }
        nextCursor:
          type: [string, "null"]
          description: Курсор наступної сторінки; null — сторінок більше немає

    Order:
      type: object
      required: [id, eventId, quantity, status, totalPrice, createdAt]
      properties:
        id:        { type: string, format: uuid }
        eventId:   { type: string, format: uuid }
        quantity:  { type: integer, minimum: 1 }
        status:
          type: string
          enum: [confirmed, cancelled]
        totalPrice: { $ref: "#/components/schemas/Money" }
        createdAt:  { type: string, format: date-time }

    CreateOrderRequest:
      type: object
      required: [eventId, quantity]
      properties:
        eventId:  { type: string, format: uuid }
        quantity: { type: integer, minimum: 1, maximum: 10 }

    CreateOrderResponse:
      description: >-
        Order плюс cancellationToken - цей секрет сервер показує лише
        один раз, у відповіді на створення. GET /orders/{orderId} його
        більше не повертає.
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          required: [cancellationToken]
          properties:
            cancellationToken: { type: string }

    CancelOrderRequest:
      type: object
      required: [cancellationToken]
      properties:
        cancellationToken: { type: string }

    Problem:
      type: object
      description: RFC 9457 Problem Details
      required: [type, title, status]
      properties:
        type:     { type: string }
        title:    { type: string }
        status:   { type: integer }
        detail:   { type: string }
        instance: { type: string }

  responses:
    BadRequest:
      description: Некоректні параметри запиту
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          examples:
            invalidLimit:
              value:
                type: /errors/invalid-query
                title: Некоректний параметр запиту
                status: 400
                detail: "Параметр limit має бути в межах 1..50"

    NotFound:
      description: Ресурс не знайдено
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          examples:
            eventNotFound:
              value:
                type: /errors/not-found
                title: Подію не знайдено
                status: 404
            orderNotFound:
              value:
                type: /errors/not-found
                title: Замовлення не знайдено
                status: 404
                detail: >-
                  Замовлення з таким id не існує, або наданий
                  cancellationToken йому не відповідає

    SeatsUnavailable:
      description: >-
        Запитана кількість місць перевищує те, що лишилось вільним на
        момент запису. Тимчасовий стан: 409, а не 410 - за хвилину хтось
        може скасувати свою бронь, і такий самий (або менший) запит пройде.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          examples:
            seatsUnavailable:
              value:
                type: /errors/seats-unavailable
                title: Недостатньо вільних місць
                status: 409
                detail: "Запитано 5 місць, вільно 2"

    EventGone:
      description: >-
        Подію скасовано організатором. На відміну від SeatsUnavailable (409) -
        це постійний стан: той самий запит ніколи не пройде, повторювати
        його немає сенсу.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          examples:
            eventCancelled:
              value:
                type: /errors/event-cancelled
                title: Подію скасовано
                status: 410
                detail: "Організатор скасував цю подію, нові замовлення неможливі"